1. The data type has been changed to BIGINT to accomodate the larger range. 2. The tenant setting is moved to the respective method where the default users being created for both the tenants on startup.

+++ src/main/resources/db/migration/V1__create_user_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE APP_USER (
-    record_id   INT AUTO_INCREMENT PRIMARY KEY,
+    record_id   BIGINT AUTO_INCREMENT PRIMARY KEY,
     first_name  VARCHAR(255),
     last_name   VARCHAR(255),
-    mobile      INT
+    mobile      BIGINT
 );
